scRNAseq harmony - fix cwl linkage and add color-data sooner
